fixed bug eliminate button

+++ controllers/mainwindow.py
@@ -109,7 +109,7 @@ class MainWindow(QMainWindow):
         #####
 
         self.ui.eliminateBtn.clicked.connect(self.eliminateCurrentRow)
-        self.ui.userTableWidget.itemSelectionChanged.connect(self.enableEliminateBtn)
+        # self.ui.userTableWidget.itemSelectionChanged.connect(self.enableEliminateBtn)
         self.ui.createBtn.clicked.connect(self.showCreateUsersDialog)
         self.ui.editBtn.clicked.connect(self.showUpdateUsersDialog)
 
@@ -236,13 +236,16 @@ class MainWindow(QMainWindow):
 
     def eliminateCurrentRow(self):
         item = self.ui.userTableWidget.currentItem()
-        user = item.data(Qt.UserRole+1)
-        user_service.delete_user(user)
-        self.loadUserTable()
-        self.ui.eliminateBtn.setEnabled(False)
+        if(item is not None):
+            user = item.data(Qt.UserRole+1)
+            user_service.delete_user(user)
+            self.loadUserTable()
+            # self.ui.eliminateBtn.setEnabled(False)
+        else:
+            print("Debe seleccionar un usuario.")
 
-    def enableEliminateBtn(self):
-        self.ui.eliminateBtn.setEnabled(True)
+    # def enableEliminateBtn(self):
+    #     self.ui.eliminateBtn.setEnabled(True)
     
     @Slot()
     def loadUserTable(self):
@@ -421,7 +424,7 @@ class MainWindow(QMainWindow):
             formUser.setLinesEditsValues(user)
             formUser.show()
         else:
-            print("Debe seleccionar un usuario")
+            print("Debe seleccionar un usuario.")
         
 
     ########################################################################
